src/requirements.txt
cryptography
Flask
requests
python-dotenv
openai==0.28.1
gunicorn
apscheduler
pymysql 
sqlalchemy